erkhet-automation/requirements.txt
playwright==1.49.1
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.32.3
undetected-chromedriver==3.5.5
selenium==4.43.0
setuptools
